glm53-kpool(copy-move,mechanical_provable): Move the buffer copy method into its sibling mixin
